python: expose Py_SetProgramName and Py_GetProgramName
Fixes sbinet/go-python#71.
